RD-4279: Relocated FirstUserJourneyButtons
