+++ widgets/deployments/src/FirstUserJourneyButtons/index.ts
@@ -1,0 +1,1 @@
+export { FirstUserJourneyButtons as default } from './FirstUserJourneyButtons';
